Specify host in default_url_options
**Why**: In production, we noticed the following 500 error:
Encoding::CompatibilityError:
incompatible character encodings: ASCII-8BIT and UTF-8

This was caused by someone setting the HTTP_HOST header to an invalid
URI, such as `¿’¿”`. By default Rails uses this HTTP_HOST header value
to generate URLs for assets and URL helpers. The HTTP_HOST header is
encoded with ASCII-8BIT, and when Rails generates the HTML to render
the page, it tries to concatenate this ASCII-8BIT host value with UTF-8
characters (the default encoding in the app), which causes the
`Encoding::CompatibilityError`.

**How**:
- Ensure the asset host is set to the server on which the app is
running. This was already done in `production.rb` but I also added it
to `development.rb`:
`config.action_controller.asset_host = Figaro.env.domain_name`
- Ensure a default host is set for URL generation. This was already done
in `production.rb` via `config.action_controller.default_url_options`,
but because we also defined a `default_url_options` method in
ApplicationController (that did not include the `host` key), it
overrode the setting in `production.rb`. I fixed this by removing the
setting from `production.rb` and setting the host in
`ApplicationController#default_url_options`.
- Don't remove the HTTP_HOST header from `env` because it prevents us
from writing a failing request spec. If the HOST header is deleted, the
test environment will provide its own valid host for URL helpers
(as opposed to letting Rails use the value from the HOST header).
- Make sure to force the header encoding to ASCII-8BIT in request specs.
Otherwise, they will be converted to UTF-8 by default.