Move WorkerHealthChecker from app/services to lib
**Why**: Middleware, such as `WorkerHealthChecker::Middleware` can't be
reloaded, but because the `WorkerHealthChecker` module was placed in
`app/services`, it automatically gets reloaded by Rails, which results
in the following error when running Sidekiq jobs while Rails is trying
to clean up and unload code:

```
ArgumentError: A copy of WorkerHealthChecker::Middleware has been
removed from the module tree but is still active!
```
The solution is to move the module to a path that is not included in
the Rails autoload paths, such as `lib`.

A negative side effect of this issue is that any background job that
ran before Rails raised the error (when it was trying to clean up and
unload the code), will keep running over and over because Sidekiq
automatically retries failed jobs. I experienced this locally by
getting the same email over and over, and I'm assuming this is the
same issue that caused some of our production users to keep getting
the same SMS over and over. Jenny also saw this in INT in June.