Whitelist origins for CORS requests
**Why**: Due to enforcing basic auth in our lower envs, we need to set
`credentials: true` for the `/api/openid_connect/token` endpoint.
However, the combination of `credentials: true` and allowing all
origins via the wildcard setting poses a security risk, and the latest
version of the `rack-cors` gem will not allow that configuration.

Until we remove basic auth, the solution is to restrict origins to
those we trust. Since the CORS endpoints are specific to OIDC clients,
I've set the origins based on the domains in
`ServiceProvider#redirect_uris`.

Reference:
https://web-in-security.blogspot.de/2017/07/cors-misconfigurations-on-large-scale.html