Unlock access key in session encryptor
**Why**: The session encryptor was not unlocking the user access keys
when it was initializing. This meant that if an old session that was
created with the old non-deterministic access key was decrypted, that
key would be used to unlock the session encryptor's user access key,
setting the random_r to the value of the random_r for the
non-deterministic key.