Prepend set_locale before action
**Why**: The set_locale before action sets the `I18n.locale` variable.
Since that is part of the global state, it needs to be set before
rendering any pages or pages will be rendered with the locale of the
previous request. Prepending the `set_locale` action means it comes
before actions that break the callback chain for things like CSRF errors
and timeouts, assuring those pages are rendered in the correct language.