Handle nil redirect_uris gracefully
**Why**: The IdP app expects the `redirect_uris` column in the
ServiceProvider table to be an array and to have a default value
of an empty array, but the dashboard app saves an empty value as
`null`, and Rails does not have a built-in mechanism for converting
the `nil` value into an empty array.

**How**: Call `#compact` on the array before iterating through the
URIs to get rid of any `nil` values since we don't care about them.