Allow configuring NameID format for each SP.
**Why**:

It's a common pattern to want to customize the NameID format used by a
Service Provider. Currently Login.gov only supports "persistent", i.e. a
UUID. But many service providers and COTS products require the NameID to
be an email address.

Login.gov claims to support both in our public SAML metadata, but this
wasn't previously true in reality.

**How**:

Add a new string column name_id_format_type to service_provider. This
string will be a key into the NAME_ID_FORMAT_TYPE_OPTIONS Hash, which
contains the information needed to pass the appropriate NameID format to

The column is nullable, so all existing SPs will continue using the
default of "persistent".

This relies on a new branch of the saml_idp gem, which adds code that
checks the principal object for a name_id_format method, similar to the
existing check for a asserted_attributes method.

In the identity-idp code, pass through name_id_format from the
ServiceProvider object at the same time that we pass through
asserted_attributes.