Allow email NameID Format on a per-SP basis
**Why**: In general, we recommend using the user's persisent UUID
as the NameID Format, but in some cases, some SPs are only able to
use the email address. The easiest way to allow those SPs to use
email address was to update the `saml_idp` gem to pick the correct
NameID format from the IdP's config based on the SP's requested
format, and then add logic in the IdP to only allow the `emailAddress`
NameID Format from whitelisted SPs.