Allow OIDC requests with Accept header "*/*"
**Why**: The `two_factor_authentication` gem includes a `before_action`
that calls `render nothing: true` if the Accept header is not set to
`html`. This means that if a user is partially signed in (i.e. with
email and password only), and then they visit a Service Provider and
make an OpenId Connect request to the IdP with an Accept header set to
`*/*`, the app will throw an `ActionView::MissingTemplate` error because
it's trying to render a view that does not exist.

`render nothing: true` is deprecated in Rails 5.1. If we submitted a PR
to the gem to replace it with `head :unauthorized`, the user would still
not be able to access the page. What we want is to allow the `*/*`
format. We could ask the gem maintainers why they only allow the `html`
format, but since we don't want the `handle_two_factor_authentication`
callback at all in some controllers (because we have our own
`user_fully_authenticated?` method for determining whether or not a user
needs 2FA), the easiest solution is to skip the callback.