Get rid of vendor validator abstraction
**Why**: The vendor validator made it difficult to clearly see where and
how the logic that commuicates with the proofer gem was. It will also be
less helpful in a world where we have multiple vendors per proofing
step. This commit tears the vendor validator down in favor of a unique
job class for each step. That job class then invokes the proofer gem
code itself.