Allow different vendors for each step in idv flow
**Why**: So that we can use different vendors to proof different kinds
of data, e.g. we can use Equifax for resolution & phone and we can use
AAMVA for state ID.

**How**: This commit adds 3 new configs to `application.yml`. Each of
these configs refers to a proofing step and specifies the vendor to be
used for that step. A caveat to this is that we removed code that
randomly selected from a set of vendors and allowed that vendor to be
used for the entire process. That should be okay since that code was
unused and growing stale.