Skip handle_two_factor_authentication
**Why**: This is a callback that is included by default in all
non-Devise controllers by the `two_factor_authentication` gem.
We don't have a need for this callback at all since in the
controllers where we do need to enforce that the user is fully
authenticated, we have our own `confirm_two_factor_authenticated`
callback.

As we've seen previously in #1799, allowing this callback to run
prevents requests with the `*/*` Accept header from succeeding.

Unfortunately, we can't skip this callback in ApplicationController
because I think it gets defined later in the chain. I am working
on a PR that will disable this callback everywhere we don't need it,
but for now, I wanted to keep this simple and disable in a controller
that shows up in our error logs.