Redirect to referer during most CSRF errors
**Why**: Signing the user out and redirecting them to the sign in
page is not helpful when they encounter a CSRF error. For example,
if they get a CSRF error while creating their password for the first
time during account creation, they might not know what they should
do to get back to the password creation screen.

Instead, we should try to redirect back to where they were so they can
try again. One exception is on the personal key screen. Due to the way
the feature works, we can't take the user back to the personal key
screen without modifying the session and regenerating a new key for
them. We don't want to make any state changes as a result of a CSRF
error for security reasons. Instead, we sign them out and redirect
to the sign in screen.

`Users::SessionsController` also has its own way of rescuing the
CSRF error because the only way I could get the flash message to
appear was to call `sign_out`, and we don't want to call `sign_out`
in `ApplicationController`. Also, in order to preserve the `request_id`
(so that users can go back to the SP), we need to use `redirect_back`,
whereas in the Personal Keys controllers, we don't want to redirect
back.

This PR also enables the Devise `unauthenticated` error message in
the scenario where a user tries to access a page that requires
authentication while the user is signed out. We capture that flash
message in analytics in order to see whether most people are getting
CSRF errors while signed in or if they try to submit a form after
their session has timed out or somehow got dropped by the browser.