Ensure load tests follow flows correctly
**Why**: In load tests, it can be difficult to be sure that the tests
are following the same path as users and are not erroring out. If
uncaught, this can lead to false positives. This commit forces the
tests to walk through each page as a user would and checks that each
page is as expected. It also provides more actionable feedback should
something go wrong.