Redirect back to SP regardless of locale
**Why**: Andrew found a bug where selecting Spanish during account
creation or authentication would not redirect you back to the SP.

The bug was in `SecureHeadersConcern`, which only allowed the redirect
to happen if the redirect URI started with the OIDC endpoint. The
problem is the URL of the OIDC endpoint is dynamic depending on the
language. When the language is not English, the SP's redirect URI no
longer starts with the same set of characters as the OIDC endpoint,
therefore the condition will always be false, and the redirect will
not be allowed.

**How**: Instead of determining whether or not to return early based on
whether or not the SP request URL is an OIDC URL, base it only on the
presence of the request URL.

I believe the original intent of the conditional in the first return
statement was for performance reasons, since SAML URLs do not need to
be added to CSP headers in the CompletionsController. The reason is
that CSP headers only matter on the last POST request before a redirect
to an external URL. In the case of OIDC, clicking Continue on the
Completions page, or after 2FA authentication, is the last form that
is submitted before the redirect back to the SP. However, for SAML,
there is an additional hidden form that is rendered and automatically
submitted via JS. This happens in SamlIdpController, which is why it
has its own CSP override.

Thanks to our use of the Null Object pattern, passing in a SAML URL
in `SecureHeadersConcern` allows the app to work as expected because
`authorize_form.valid?` will return `false`.