Don't use Devise's stored location
**Why**: When a user tries to access a page that requires
authentication (such as their account page) while they are signed out,
Devise automatically saves that URL as a "stored location", and when
the user signs in, it takes them to that page. From what I can tell,
this is only as a convenience to the user. For example, their session
might have timed out, but their browser still shows a form that
requires authentication. If they submit the form, they will be
prompted to sign in, and once they do, they will be taken back to the
form so they can continue what they were doing.

In our app, this Devise feature can cause more problems than it is
trying to solve. For example, if a user tries to visit their account
page while signed out, they will be prompted to sign in. If they then
visit a Service Provider and initiate login, after they sign in, they
will end up on their account page instead of back at the SP. This is
because in `after_sign_in_path_for(user)`, the order of lookup is the
stored location, then the SP request URL. The reason for this lookup
order is to support the reauthentication feature, where a user is
prompted to enter their current password and go through 2FA when they
are attempting to change their email, password, or phone (if enough
time has passed since they last authenticated). For example, if they
were trying to update their email, we store the location of the email
update page so that we can take them there after they go through the
reauthentication.

**How**: In our app, there aren't that many pages a user would need to
access while authenticated. The account page is basically the only one.
Moreover, we have callbacks on all controllers that will take the user
to where they need to be. In other words, if a user loses their place
during account creation and they sign out or their session times out,
they will always end up back where they left off. This feature doesn't
depend on Devise.

- Don't call `stored_location_for(user)`
- In `ReauthnRequiredController`, don't use Devise's method for
storing the location. Instead, use our own custom method and store
it in our own custom session key.
- Remove unnecessary location storage in `SamlIdpController`
- Add Devise's stored location to analytics in
`Users::SessionsController` to see how often people attempt to visit
pages before signing in. I have a theory that this is the cause of some
production errors I have seen.