Validate phone for Voice OTP delivery preference
**Why**: We noticed Twilio errors due to people attempting to make
phone calls to India, a country that we currently don't support.
While using the app normally, it is not possible to make phone calls
to unsupported countries or area codes. For example, when setting up
your 2FA phone for the first time, we have both client and server
validation that prevents making phone calls to unsupported countries
and area codes. In addition, once you confirm your 2FA number via SMS,
and then sign in, if your phone number is from an unsupported country
or area code, the link to get a phone call does not appear.

However, given that many of our SMS messages to India are being blocked,
it's not surprising that people are trying other means. Some clever
folks noticed that the OTP delivery method is part of the URL on the
page that prompts for your OTP (`/login/two_factor/sms`), and they
changed `sms` to `voice`, at which point the link to get a phone call
was visible, and they clicked it hoping to get a phone call. This also
resulted in their default OTP delivery preference being set to Voice,
and every time they sign in, we try to send them a phone call, which
increases the amount of Twilio errors we get.

**How**:
- Instead of automatically sending an OTP using the user's current
otp_delivery_preference, first validate the preference. If the
preference is voice, but the number is unsupported, then we send the
OTP via SMS and redirect to `login/two_factor/sms`. Additionally,
if the current preference is voice, we make sure to reset it to 'sms'.

- If a user tries to manipulate the URL to access the voice page,
we check to see if they are allowed to access the voice page, and if
not, we redirect to the SMS page

- If a user tries to manipulate the URL to send an OTP via voice,
we validate the preference, and if it fails we redirect to the SMS page