Log email domain name during registration
**Why**: I noticed a few thousand errors over the past month due
to people entering invalid emails when registering. Since most modern
browsers have built-in validations for the email format (i.e. it has to
have some characters, then an @ character, then a domain name), the
error is most likely an invalid domain name, which we validate using
the valid_email gem, which performs MX validation.

We have also noticed that the MX validation has been timing out for
some users, and we suggested disabling it. However, before we do that,
I'd like to get some data on these errors to determine whether or not
they are due to MX validation failing, and to see if MX validation is
worth keeping.