Update rate limit settings
**Why**: We want to be able to rate limit email+IP combinations
separately from generic logins per IP. We also want to set the req/ip
and logins/ip rate limits to track only mode for now until we get
enough data from USAJOBS customers to figure out the optimal limit and
period settings. Finally, we want to update the 429 page with a more
helpful message.

**How**:
- Add Figaro config settings to allow us to easily toggle between track
and throttle mode for the req/ip and logins/ip rate limits.
- Add a blocklist with Allow2Ban to limit logins per email+IP. We could
have used a throttle, but a blocklist provides more control over the
bantime, which can be separate from the findtime. It also makes the
bantime more accurate. In a throttle, the cache key gets cleared at
the beginning of the minute. For example, if the throttle was set to
5 requests per 60 seconds, if you made 6 requests between 11:00:50 and
11:00:55, you be rate limited only until 11:01:00, at which point you
have another 5 tries until 11:02:00, giving you 10 tries over 70
seconds.
- Log rate limit triggers via Analytics as opposed to Rails.logger
so that we benefit from the request and user_agent parsing. Perhaps more
importantly, adding the request host to the hash we were passing to
Rails.logger resulted in a failing spec in `headers_spec` because the
`to_json` call was not able to convert the ASCII-8BIT encoded host.
- Update the text on the 429 page to make it more helpful to the user.