Ensure user can view their new personal key
**Why**: We have received reports from users who were no longer able to
sign in with their personal key because they no longer had one. This
was very easy to reproduce as follows:
1. Sign in with your email and password
2. On the 2FA prompt page, click the link to sign in with your personal
key
3. After you type in your personal key, press the return button on your
keyboard or click the Submit button multiple times really fast

The result was that the user was redirected somewhere other than the
personal key page, and because the PersonalKeyForm reset the personal
key to `nil` as soon as the user entered the correct key in step 3, and
because the new key does not get generated until the personal key view
is rendered, the user ended up with a nil personal key and was no
longer able to sign back in with their key (because the link to sign in
with a personal key does not appear if you don't have a personal key).

**How**:
- Never set a user's personal key to nil
- Generate the new key in `PersonalKeyVerificationController` and store
it in the session, so that if the user double-submits a form by mistake,
we still have access to their key and can redirect them to the
PersonalKey page if they get redirected to the account page and have not
yet acknowledged their new key.
- Store the key in the `user_session`, not global `session`.
- Track visits to `PersonalKeyVerificationController` and
`PersonalKeysController` to make it easier to see step by step what
happened to the user.
- Write tests for both SAML and OIDC using real-life scenarios for
better end-to-end tests.