Set a timeout for MX validation
**Why**: It's a best practice in general to always set timeouts for any
external connection. Without a timeout for MX validation, the app will
sit waiting for a response for more than a minute by default, which can
queue up requests and bring the server down. We use Rack::Timeout to
stop any long-running requests (currently set to 30 seconds, but I have
a PR to halve that), but that's not meant as a solution for
long-running requests. Rack::Timeout is only supposed to be used as a
debugging tool to see where we have long-running requests,
and then our goal is to make them faster or introduce timeouts.

If MX validation takes more than 2 or 3 seconds, we should stop it
immediately and let the user try again later if the domain name was
indeed valid.