Don't discard user access keys in SessionEncryptor
**Why**: The SessionEncryptor memoized an initialized UserAccessKey to
keep from recomputing scrypt hashes, but did not memoize a made or
unlocked access key. The reason for this is that different instances
share user access keys with different encryption keys, and locking
/ unlocking saves a read-only access key. This meant that every time a
session was loaded or saved a KMS operation needed to be invoked to
compute the CEK from the encryption key.

This commit changes the session encryptor to memoize an encryption key,
and memoize a map of decryption keys for encrypting sessions saved by
other instances. This means that the instance only has to call KMS to
compute the CEK for a user access key the first time it encrypts a
session with its key, and the first times it decrypts a session
encrypted with a different instances keys.