Don't redirect to SP when already signed in
**Why**:
Consider the following scenario:
1. Visit IdP via SP
2. Start account creation in tab 1 up until you enter your email address
3. Click the confirmation link in your email, which will most likely
open a new tab in your browser. Tab 2 is where you will continue your
account creation
4. Go all the way through until the agency handoff page but don't click
Continue just yet
5. Go back to tab 1 and refresh the page. This will call
`SignUp::EmailsController#show`, which will redirect to
`sign_up_email_url` because `session[:email]` is blank. In turn,
the `/sign_up/enter_email` page calls the `require_no_authentication`
callback because we restrict access to that page to users who are not
signed in. If a user who is already signed in tries to visit the
email registration page, we were redirecting them to
`after_sign_in_path_for`, which in this case will be
`sp_session[:request_url]`
6. Go back to tab 2 and click "Continue" on the agency handoff page.
In the case of USAJOBS, this will result in an error because they don't
allow the same request to be made if you're already signed in.

Or this scenario:
1. Visit IdP via SP
2. Right-click on the "Sign in" button and open it in a new tab
3. In the new tab, sign in and 2FA, which will redirect you back to the
SP.
4. Go back to the old tab (which should be on the `sign_up/start` page)
and refresh it. Before the fix in this PR, this would have taken you
back to the SP with an error if the SP is USAJOBS.

**How**:
- Redirect to `signed_in_url` if already signed in and trying to access
pages that are meant for signed out users. This is consistent with
`check_user_needs_redirect` in `Users::SessionsController`.
- To prevent the "For your security, we clear what you entered" flash
message from appearing on the account page, we only display it if a
current user is not present, since that message is only meant for
signed out users due to the automatic page refresh via JS (to prevent
a CSRF error when they try to submit the form after the session has
timed out).