Change OTP timestamp column to integer
**Why**: In the `User#authenticate_totp` method in the two factor
authentication gem, the value assigned to `User#totp_timestamp` is an
integer, but the column type was datetime. This meant that the value was
saved as nil, preventing the logic to prevent TOTP re-use from working.

This commit also adds a test to verify that TOTP re-use is not possible.