Don't allow email as password
**Why**: Allowing someone to use their email as their password (which
the zxcvbn algorithm will not flag as a weak password by default)
reduces the security of their account.

**How**: Add a small dictionary of words that the password strength
meter can look up to adjust the calculated strength. The words are
the user's email, the different word components of the user's email,
and "login.gov". Note that the Dropbox algorithm does not outright ban
a password just because it contains a combination of those words. It
just can't be exactly one of those words, and that word only. In other
words, if the user's email is "test@test.com", you can set a password
that is "test@test.com login.gov", but you can't use a password that
is "test@testcom" or "login.gov".

Note that the ability to set a password that contains both the user's
email and "login.gov" is due to our minimum password strength being set
to 3. If we increase it to 4, a user won't be able to set such a
password.