Log Pii::EncryptionError during sign in
**Why**: We've had reports of people who were not able to sign in
with their email and password, even after resetting their password
multiple times. If the app raises `Pii::EncryptionError`, then
`valid_password?` will be false and the user will see the "invalid email
or password" message. Encryption errors are expected when the password
is indeed wrong, but we'd like to know if there are other errors being
raised. To help troubleshoot this, we will log the error messages.