Add a build script for deploy automation.
**Why**:

Previously these deploy steps were performed by chef cookbooks in the
identity-devops repo. This was an abstraction layer violation because it
required identity-devops to be aware of the internals of how
identity-idp builds and bootstrapps the application, which made it more
difficult to make changes.

**How**:

Move several build steps into a build script in the identity-idp repo,
which is called at boostrapping time by identity-devops cookbooks.