Move assets:precompile into separate script.
**Why**:

The rake assets:precompile process has to be run after deploy/activate
because it relies on the full database configuration being available.

We still want to run as many of the build steps earlier so that they can
be part of a static deploy artifact in a future build system.

**How**:

Create a separate build-post-config script. The full order of
bootstrapping will now be:

- deploy/build
- deploy/activate
- deploy/build-post-config