Remove references to `yarn build`
**Why**: Literally, all it does is run the `true` command. The `scripts`
section of `package.json` is meant to specify certain scripts that can
be run. Each script has a name, which is the key of the hash, and the
value is meant to contain an actual script that can be run on the
command line.

Reference: https://yarnpkg.com/en/docs/cli/run