Memoize UAKs on EncryptedAttribute class
**Why**: Initializing a user access key means calculating a SCrypt hash
which is expensive. If we've already calculated the SCrypt hash for a
given key/cost, we shouldn't have to again. The EncryptedAttribute class
was initializing a UAK with a shared key each time it was called for a
new model. This commit memoizes a UAK when we initialize it for a key so
we are not constantly re-initializing it.