Validate phone number pattern before submission
**Why**: We want to be able to leverage several phone validation
options, so we are (1) moving the validation to the backend, (2)
giving feedback to the user when the phone number is considered
valid, (3) enabling "Send Code" only when the phone number is
considered valid.

**How**: We send a POST to an endpoint when the phone input field blurs
or the country code changes. The result of the POST indicates if the
number is valid. We can add more information later.