Use SecureRandom.urlsafe_base64 for request ids
**Why**: So that URLs are more compact and request IDs are not
guessable. The UUID spec specifies that UUIDs should not be assumed to
be unguessable.

Currently, if someone guesses a UUID, they can hijack a user's service
provider request. That means they can break the sign in flow
for that user, but can't sign in as that user. It would be a concern
if in the future a user sets some kind of state at an SP and then signs
in.