Deploy RC 53 to staging (#2050)
* Add rake task to create readonly database user

**Why**: So we can create a readonly database user to be used in the
console after #1996

* Fall back to A record if MX record doesn't exist

**Why**: Some email servers don't configure an MX record and rely on
senders falling back to the A record, as recommended by the RFC.

**How**: The email validation gem takes a parameter to do a check of
the MX record and fall back to an A record if there is no MX record.

* assume some jitter in timing when comparing time differences

* LG-132 Updated OTP and Session timeouts (#2041)

* Updated OTP and Session timeouts

**Why**: In order to provide a bit more time for USAJobs users.

* LG-132 Fixed tests

**Why**: Changes to config broke a couple of them.

* Select AWS SES region from a pool

**Why**: We need to steadily increase our sending volume when moving
between AWS regions to keep from being block as spam. This commit allows
us to configure what percentage of mail gets sent from what region.

* Validate phone number pattern before submission

**Why**: We want to be able to leverage several phone validation
options, so we are (1) moving the validation to the backend, (2)
giving feedback to the user when the phone number is considered
valid, (3) enabling "Send Code" only when the phone number is
considered valid.

**How**: We send a POST to an endpoint when the phone input field blurs
or the country code changes. The result of the POST indicates if the
number is valid. We can add more information later.

* Validate Twilio requests to /api/voice/otp

**Why**: To make sure that only requests that contain Twilio's signature
are allowed.

Reference:
https://twilio-ruby.readthedocs.io/en/latest/usage/validation.html#rack-middleware

* LG-123 Show OTP expiration to user on web page

**Why**: So they know up front when it will expire.

* LG-1 delete account: spanish and french translations (#2042)

**Why**: Translations are not available for French and Spanish

**How** Update the locale files