Refactor deploy/activate script
**Why**: If the `deploy/activate` script is a shell script, it requires
calling `Bundler.setup` so that the files that the script depends on
are loaded. The problem is that the `Bundler.setup` call was preventing
us from running parallel builds in Circle CI.

I removed the `Bundler.setup` call in #2060 to see if that would allow
parallel builds, which it did, but I forgot to check if this would break
build deployment, and it did.

In order to allow the script to run during deployment and also allow
parallel builds to work, I defined a shell script that will fetch the
config from S3 by calling the ruby file directly using `bundle exec`.
That way, the gems will already be loaded. This also separates concerns
and makes it easier to test the S3 fetching in isolation. It also keeps
things consistent with the other scripts in the `deploy` directory.