Update Rails setup commands
**WHY**:
Onboarding and setup fails due to an issue with the `db:schema:load`
rake task that sets the "environment" of the development database to
"test", preventing subsequent rake tasks from completing successfully.

**HOW**:
Using the `db:migrate:reset` rake task instead of `db:reset` which
uses `db:migrate` instead of `db:schema:load` to prepare the database.
The downside is that this is slower since it runs all migrations and
potentially more error-prone if the migrations somehow get out of sync
with the schema. Ideally, the issue should be investigated and
possibly a PR submitted to Rails to actually fix the issue.