Use shared examples to test cancelation behavior
**Why**: So that we can test that cancelation is consistent across IdV
steps.

Note that this tests the behavior with the different protocols because
the difference between cancellation behavior with and without an SP is
surprisingly different.

Also, this commit adds some empty contexts for steps that don't have a
cancel button with the assumption that we want to add a cancel button to
those steps.