LG-211 Allow looking up users via web interface
**Why**: When troubleshooting issues reported by users, it is
cumbersome to have to connect to the VPN, then SSH into the servers,
then look up a user in the Rails console, or look at their events in
Kibana. It would be much faster and easier to look them up via the
website.

**How**: Add a page where privileged users can look up users and see
pertinent information related to their account. Only users who are
flagged with the `admin` or `tech` role can access that page.