LG-244 Exclude events with invalid tokens
**Why**:
Ahoy uses UUIDs to track visits and visitors. To determine the
value of `visit_id` and `visitor_id`, Ahoy first checks the request's
headers, then cookies, and if neither are present, it generates a new
random UUID.

In version 1.x of the Ahoy gem, if a request was made with headers
and/or cookies containing non-UUID bogus values, Ahoy would
automatically convert them to a valid UUID. This masked the fact that we
get events that should actually be ignored, such as when pentesters make
those fake requests.

In version 2.x of the gem (which we upgraded to recently), the tokens
are still altered (by removing anything that's not an alphanumeric
character or a dash, and limiting the length to 64 characters) but not
converted to a UUID, which then exposed the fact that some people are
passing in non-UUID values in headers and/or cookies. We want to ignore
these invalid visits so they don't clutter our analytics logs.