Don't check worker health if we have no workers.
Also add an "all_checks_healthy" key to the multi checker output, which
makes it easier to grep for success in clients like New Relic that don't
really speak JSON.

**Why**:

If we switch the activejob queue adapter to inline or async, then it
doesn't make sense for the top level health check to test redis queue
health because they won't be used.

**How**:

If the activejob queue adapter is set to inline or async, remove the
worker check from the multi health checker that serves the `/api/health`
endpoint so that health checks will still pass if sidekiq is not in use.