Create purpose-built encryptor for encrypting PII
**Why**: To consolidate the logic that is used to encrypt PII and to
prevent code that invokes the encryption logic from having to manage
user access keys. This consolidates things into a simpler encryption API
that makes things like migrating to a new algorithm easier.

Note that there is a performance hit when encrypting / decrypting PII
since we have to calculate 2 SCrypt hashes now, one to verify the
password and one to encrypt/decrypt the data. This is an intentional
tradeoff to make the API for the encryption module simpler with the
understanding that it is very likely that the logic used to verify the
password and encrypt PII will diverge in the future.