LG-246 Create attribute encryptor (#2155)
**Why**: To prevent the models with encrypted attributes from having to
worry about managing user access keys or key rotation. This commit
starts the move to a place where user access key management is handled
by encryptors and not by the clients using those encryptors. This will
make it easier to migrate to a different encryption solution if/when
that becomes necessary.