LG-273 Sanitize Ahoy headers and cookies
**Why**: Ahoy headers and cookies with invalid UTF-8 bytes should not
cause the app to throw a 500 error. The root cause of the exception is
in the Ahoy gem's code, which manipulates the header/cookie value to
remove certain strings using `gsub`.

**How**: Capture invalid headers and cookies in our utf8_sanitizer class
before they get to the Ahoy gem. Ideally, the Ahoy gem would remove the
token manipulation and instead, allow clients to validate tokens
however they please. I proposed this to the gem maintainer and am
waiting to hear back.