Lg 248 state id screen (#2152)
Verify the user has a state id from a supported jurisdiction before requiring them to provide the rest of the profile attributes. Provide feedback and allow them to exit the flow. This also replaces the screen that informed the user of ALL the info they need to continue that was deemed unnecessary.