Disable recaptcha by default in production.
**Why**:

We have a large number of different testing environments. When we add
new functionality to the idp that requires additions to the secrets
config, it's most friendly to not enable that functionality in
environments where we haven't also added the corresponding secrets.

**How**:

Set recaptcha to 0 by default. This should be overridden by app secrets
in environments that wish to enable recaptcha.

Ideally we would have a private version controlled place to set per-env
configuration, but for now app secrets is our only way.

In addition, it could be very surprising and unintuitive that a file
named `application.yml.example` is actually used as the basis for
production configuration, so add some comments that
application.yml.example is used in prod.